+++ tests/util/base_sort_test_case.rs
@@ -47,7 +47,7 @@ pub trait BaseSortTestCase {
         let arr_len = arr.len();
         {
             let mut sorter = self.new_sorter(random, &mut to_sort);
-            let result = sorter.sort(o, o + arr_len);
+            let result = sorter.sort(o as i32, (o + arr_len) as i32);
             assert!(result.is_ok());
         }
         self.assert_sorted(&mut arr, &to_sort[o..o + arr_len]);
